scripts/requirements.txt
anthropic>=0.40.0
snowflake-connector-python>=3.6.0
pandas>=2.0.0
python-dotenv>=1.0.0
